# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/ops/event/curEvent": {
            "get": {
                "description": "获取当前活跃的告警事件列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警事件"
                ],
                "summary": "获取当前告警事件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenant_id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "告警级别",
                        "name": "alert_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警名称",
                        "name": "alert_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "实例",
                        "name": "instance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "服务名称",
                        "name": "service_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.AlertCurEvent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/ops/event/hisEvent": {
            "get": {
                "description": "获取历史告警事件列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警事件"
                ],
                "summary": "获取历史告警事件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenant_id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "告警级别",
                        "name": "alert_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警名称",
                        "name": "alert_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "实例",
                        "name": "instance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "服务名称",
                        "name": "service_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.AlertHisEvent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AlertCurEvent": {
            "type": "object",
            "properties": {
                "alarmAggregation": {
                    "type": "boolean"
                },
                "annotations": {
                    "type": "string"
                },
                "datasource_id": {
                    "type": "string"
                },
                "datasource_type": {
                    "type": "string"
                },
                "duty_user": {
                    "type": "string"
                },
                "effectiveTime": {
                    "$ref": "#/definitions/models.EffectiveTime"
                },
                "eval_interval": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "first_trigger_time": {
                    "description": "第一次触发时间",
                    "type": "integer"
                },
                "first_trigger_time_format": {
                    "type": "string"
                },
                "for_duration": {
                    "type": "integer"
                },
                "is_recovered": {
                    "type": "boolean"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "last_eval_time": {
                    "description": "上一次评估时间",
                    "type": "integer"
                },
                "last_send_time": {
                    "description": "上一次发送时间",
                    "type": "integer"
                },
                "metric": {
                    "type": "object",
                    "additionalProperties": true
                },
                "noticeGroup": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "notice_id": {
                    "description": "默认通知对象ID",
                    "type": "string"
                },
                "recoverNotify": {
                    "type": "boolean"
                },
                "recover_time": {
                    "description": "恢复时间",
                    "type": "integer"
                },
                "recover_time_format": {
                    "type": "string"
                },
                "repeat_notice_interval": {
                    "description": "重复通知间隔时间",
                    "type": "integer"
                },
                "response_time": {
                    "type": "string"
                },
                "rule_id": {
                    "type": "string"
                },
                "rule_name": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                },
                "time_remaining": {
                    "type": "integer"
                }
            }
        },
        "models.AlertHisEvent": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "string"
                },
                "datasource_id": {
                    "type": "string"
                },
                "datasource_type": {
                    "type": "string"
                },
                "eval_interval": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "first_trigger_time": {
                    "description": "第一次触发时间",
                    "type": "integer"
                },
                "is_recovered": {
                    "type": "boolean"
                },
                "last_eval_time": {
                    "description": "最近评估时间",
                    "type": "integer"
                },
                "last_send_time": {
                    "description": "最近发送时间",
                    "type": "integer"
                },
                "metric": {
                    "type": "object",
                    "additionalProperties": true
                },
                "recover_time": {
                    "description": "恢复时间",
                    "type": "integer"
                },
                "rule_id": {
                    "type": "string"
                },
                "rule_name": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "tenantId": {
                    "type": "string"
                }
            }
        },
...
        "models.EffectiveTime": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "integer"
                },
                "week": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
    }
}